# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Final Project Backend API",
    "version": "0.1.0",
    "description": "API documentation for mobile and web clients. Most protected endpoints require a Supabase access token via Authorization: Bearer <token>."
  },
  "servers": [
    {
      "url": "https://www.medi-buddy.xyz",
      "description": "Production Server (HTTPS)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase access token"
      }
    },
    "schemas": {
...
      "AuthV2RefreshRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "AuthV2RefreshResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          }
        },
        "required": [
          "accessToken"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "MedicineType": {
        "type": "string",
        "enum": [
          "ORAL",
          "TOPICAL"
        ]
      },
...
      "MedicineListMedicineInfo": {
        "type": "object",
        "properties": {
          "mediId": {
            "type": "integer"
          },
          "mediThName": {
            "type": "string"
          },
          "mediEnName": {
            "type": "string"
          },
          "mediTradeName": {
            "type": "string",
            "nullable": true
          },
          "mediType": {
            "$ref": "#/components/schemas/MedicineType"
          },
          "mediPicture": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "mediId",
          "mediThName",
          "mediEnName",
          "mediType"
        ]
      },
...
      "ScheduleType": {
        "type": "string",
        "enum": [
          "DAILY",
          "WEEKLY",
          "INTERVAL",
          "CYCLE"
        ]
      },
      "MealRelation": {
        "type": "string",
        "enum": [
          "BEFORE_MEAL",
          "AFTER_MEAL",
          "WITH_MEAL",
          "NONE"
        ]
      },
      "DayOfWeek": {
        "type": "string",
        "enum": [
          "SUN",
          "MON",
          "TUE",
          "WED",
          "THU",
          "FRI",
          "SAT"
        ],
        "description": "Day of week values (SUN=Sunday ... SAT=Saturday)"
      },
      "MedicineRegimenTime": {
        "type": "object",
        "properties": {
          "timeId": {
            "type": "integer"
          },
          "time": {
            "type": "string",
            "description": "Time in HH:MM format (e.g., '09:00')"
          },
          "dose": {
            "type": "integer",
            "minimum": 1
          },
          "unit": {
            "type": "string"
          },
          "mealRelation": {
            "$ref": "#/components/schemas/MealRelation"
          },
          "mealOffsetMin": {
            "type": "integer",
            "nullable": true,
            "description": "Required when mealRelation is not NONE; must be null/omitted when mealRelation is NONE"
          }
        },
        "required": [
          "timeId",
          "time",
          "dose",
          "unit",
          "mealRelation"
        ]
      },
      "MedicineRegimenTimeInput": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "time": {
            "type": "string",
            "description": "Time in HH:MM format (00:00-23:59)"
          },
          "dose": {
            "type": "integer",
            "minimum": 1
          },
          "unit": {
            "type": "string"
          },
          "mealRelation": {
            "$ref": "#/components/schemas/MealRelation"
          },
          "mealOffsetMin": {
            "type": "integer",
            "nullable": true,
            "minimum": 0,
            "description": "Required when mealRelation is not NONE; must be null/omitted when mealRelation is NONE"
          }
        },
        "required": [
          "time",
          "dose",
          "unit",
          "mealRelation"
        ]
      },
      "MedicineRegimenCreateDailyRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "mediListId": {
            "type": "integer"
          },
          "scheduleType": {
            "type": "string",
            "enum": [
              "DAILY"
            ]
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "description": "Required for DAILY schedule"
          },
          "times": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicineRegimenTimeInput"
            },
            "minItems": 1
          }
        },
        "required": [
          "mediListId",
          "scheduleType",
          "startDate",
          "endDate",
          "times"
        ]
      },
      "MedicineRegimenCreateWeeklyRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "mediListId": {
            "type": "integer"
          },
          "scheduleType": {
            "type": "string",
            "enum": [
              "WEEKLY"
            ]
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "daysOfWeek": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DayOfWeek"
            },
            "minItems": 1,
            "uniqueItems": true,
            "description": "Required for WEEKLY schedule."
          },
          "times": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicineRegimenTimeInput"
            },
            "minItems": 1
          }
        },
        "required": [
          "mediListId",
          "scheduleType",
          "startDate",
          "daysOfWeek",
          "times"
        ]
      },
      "MedicineRegimenCreateIntervalRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "mediListId": {
            "type": "integer"
          },
          "scheduleType": {
            "type": "string",
            "enum": [
              "INTERVAL"
            ]
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "intervalDays": {
            "type": "integer",
            "minimum": 1,
            "description": "Required for INTERVAL schedule."
          },
          "times": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicineRegimenTimeInput"
            },
            "minItems": 1
          }
        },
        "required": [
          "mediListId",
          "scheduleType",
          "startDate",
          "intervalDays",
          "times"
        ]
      },
      "MedicineRegimenCreateCycleRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "mediListId": {
            "type": "integer"
          },
          "scheduleType": {
            "type": "string",
            "enum": [
              "CYCLE"
            ]
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cycleOnDays": {
            "type": "integer",
            "minimum": 1,
            "description": "Required for CYCLE schedule."
          },
          "cycleBreakDays": {
            "type": "integer",
            "minimum": 1,
            "description": "Required for CYCLE schedule."
          },
          "times": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicineRegimenTimeInput"
            },
            "minItems": 1
          }
        },
        "required": [
          "mediListId",
          "scheduleType",
          "startDate",
          "cycleOnDays",
          "cycleBreakDays",
          "times"
        ]
      },
      "MedicineRegimenCreateRequest": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/MedicineRegimenCreateDailyRequest"
          },
          {
            "$ref": "#/components/schemas/MedicineRegimenCreateWeeklyRequest"
          },
          {
            "$ref": "#/components/schemas/MedicineRegimenCreateIntervalRequest"
          },
          {
            "$ref": "#/components/schemas/MedicineRegimenCreateCycleRequest"
          }
        ],
        "discriminator": {
          "propertyName": "scheduleType",
          "mapping": {
            "DAILY": "#/components/schemas/MedicineRegimenCreateDailyRequest",
            "WEEKLY": "#/components/schemas/MedicineRegimenCreateWeeklyRequest",
            "INTERVAL": "#/components/schemas/MedicineRegimenCreateIntervalRequest",
            "CYCLE": "#/components/schemas/MedicineRegimenCreateCycleRequest"
          }
        }
      },
      "MedicineRegimenItem": {
        "type": "object",
        "properties": {
          "mediRegimenId": {
            "type": "integer"
          },
          "mediListId": {
            "type": "integer",
            "nullable": true
          },
          "scheduleType": {
            "$ref": "#/components/schemas/ScheduleType"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "nextOccurrenceAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "daysOfWeek": {
            "type": "string",
            "nullable": true
          },
          "intervalDays": {
            "type": "integer",
            "nullable": true
          },
          "cycleOnDays": {
            "type": "integer",
            "nullable": true
          },
          "cycleBreakDays": {
            "type": "integer",
            "nullable": true
          },
          "medicineList": {
            "type": "object",
            "nullable": true,
            "properties": {
              "mediListId": {
                "type": "integer"
              },
              "mediNickname": {
                "type": "string",
                "nullable": true
              },
              "pictureOption": {
                "type": "string",
                "nullable": true
              },
              "medicine": {
                "$ref": "#/components/schemas/MedicineListMedicineInfo"
              }
            }
          },
          "times": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicineRegimenTime"
            }
          }
        },
        "required": [
          "mediRegimenId",
          "scheduleType",
          "startDate",
          "times"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/auth/v2/refresh": {
      "post": {
        "tags": [
          "Auth (V2)"
        ],
        "summary": "Refresh access token",
        "description": "Exchange a refresh token for a new access token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthV2RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthV2RefreshResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mobile/v1/medicine-regimen/create": {
      "post": {
        "tags": [
          "Mobile - MedicineRegimen"
        ],
        "summary": "Create medicine regimen",
        "description": "Create a new medicine regimen with schedule settings and multiple times. Schedule type determines which fields are required: DAILY requires endDate, WEEKLY requires daysOfWeek, INTERVAL requires intervalDays, CYCLE requires cycleOnDays and cycleBreakDays. Fields not applicable to the selected scheduleType are rejected. Each time must have mealOffsetMin when mealRelation is not NONE (and must be null/omitted when mealRelation is NONE).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicineRegimenCreateRequest"
              },
              "examples": {
                "DAILY": {
                  "summary": "DAILY (requires endDate)",
                  "value": {
                    "mediListId": 123,
                    "scheduleType": "DAILY",
                    "startDate": "2026-01-01T00:00:00.000Z",
                    "endDate": "2026-12-31T00:00:00.000Z",
                    "times": [
                      {
                        "time": "09:00",
                        "dose": 1,
                        "unit": "tablet",
                        "mealRelation": "BEFORE_MEAL",
                        "mealOffsetMin": 30
                      }
                    ]
                  }
                },
                "WEEKLY": {
                  "summary": "WEEKLY (pick daysOfWeek from dropdown)",
                  "value": {
                    "mediListId": 123,
                    "scheduleType": "WEEKLY",
                    "startDate": "2026-01-01T00:00:00.000Z",
                    "endDate": null,
                    "daysOfWeek": [
                      "MON",
                      "WED",
                      "FRI"
                    ],
                    "times": [
                      {
                        "time": "09:00",
                        "dose": 1,
                        "unit": "tablet",
                        "mealRelation": "NONE"
                      }
                    ]
                  }
                },
                "INTERVAL": {
                  "summary": "INTERVAL (every N days)",
                  "value": {
                    "mediListId": 123,
                    "scheduleType": "INTERVAL",
                    "startDate": "2026-01-01T00:00:00.000Z",
                    "endDate": null,
                    "intervalDays": 2,
                    "times": [
                      {
                        "time": "09:00",
                        "dose": 1,
                        "unit": "tablet",
                        "mealRelation": "WITH_MEAL",
                        "mealOffsetMin": 0
                      }
                    ]
                  }
                },
                "CYCLE": {
                  "summary": "CYCLE (on/off pattern)",
                  "value": {
                    "mediListId": 123,
                    "scheduleType": "CYCLE",
                    "startDate": "2026-01-01T00:00:00.000Z",
                    "endDate": null,
                    "cycleOnDays": 7,
                    "cycleBreakDays": 3,
                    "times": [
                      {
                        "time": "09:00",
                        "dose": 1,
                        "unit": "tablet",
                        "mealRelation": "AFTER_MEAL",
                        "mealOffsetMin": 30
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicineRegimenItem"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed to create regimen for this medicine list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Medicine list not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}